requirements.txt
Keras==2.2.4
pandas==0.25.0
numpy==1.17.0
scikit-learn==0.21.3
category_encoders==2.0.0
mlxtend==0.17.0
matplotlib==3.1.1
seaborn==0.9.0
scipy==1.3.1
ipykernel==5.1.2
feather-format==0.4.0
eli5==0.9.0
tensorflow
plotly
lightgbm
scikit-optimize
catboost
jupyter_contrib_nbextensions
